+++ tests/testthat/fixtures/desc/quoted-own-function-name/R/do-good-things.R
@@ -1,0 +1,3 @@
+do_good_things <- function(x) {
+  x
+}